--- a/reco_utils/evaluation/python_evaluation.py
+++ b/reco_utils/evaluation/python_evaluation.py
@@ -10,7 +10,7 @@ from sklearn.metrics import (
     r2_score,
     explained_variance_score,
     roc_auc_score,
-    log_loss
+    log_loss,
 )
 
 from reco_utils.common.constants import (
@@ -31,6 +31,7 @@ def check_column_dtypes(f):
         1. whether the input columns exist in the input dataframes.
         2. whether the data types of col_user as well as col_item are matched in the two input dataframes.
     """
+
     @wraps(f)
     def check_column_dtypes_wrapper(
         rating_true,
@@ -77,11 +78,17 @@ def check_column_dtypes(f):
             *args,
             **kwargs
         )
+
     return check_column_dtypes_wrapper
 
 
 def merge_rating_true_pred(
-    rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
+    rating_true,
+    rating_pred,
+    col_user=DEFAULT_USER_COL,
+    col_item=DEFAULT_ITEM_COL,
+    col_rating=DEFAULT_RATING_COL,
+    col_prediction=PREDICTION_COL,
 ):
     """Join truth and prediction data frames on userID and itemID
     
@@ -95,30 +102,25 @@ def merge_rating_true_pred(
 
     Returns:
         pd.DataFrame: Merged pd.DataFrame
+        str: Column with the true ratings
+        str: Column with the predicted ratings
     """
-    # Select the columns needed for evaluations
-    rating_true = rating_true[[col_user, col_item, col_rating]]
-    rating_pred = rating_pred[[col_user, col_item, col_prediction]]
-
-    if col_rating == col_prediction:
-        rating_true_pred = pd.merge(
+    suffixes = ["_true", "_pred"]
+    # Apart from merging both dataframes, pd.merge will rename the columns with the suffixes only if the rating
+    # column name of rating_true is the same as the name rating column name in rating_pred
+    rating_true_pred = pd.merge(
             rating_true,
             rating_pred,
             on=[col_user, col_item],
-            suffixes=["_true", "_pred"],
+            suffixes=suffixes,
         )
-        rating_true_pred.rename(
-            columns={col_rating + "_true": DEFAULT_RATING_COL}, inplace=True
-        )
-        rating_true_pred.rename(
-            columns={col_prediction + "_pred": PREDICTION_COL}, inplace=True
-        )
+    if col_rating == col_prediction:
+        column_select_true = col_rating + suffixes[0]
+        column_select_pred = col_prediction + suffixes[1]
     else:
-        rating_true_pred = pd.merge(rating_true, rating_pred, on=[col_user, col_item])
-        rating_true_pred.rename(columns={col_rating: DEFAULT_RATING_COL}, inplace=True)
-        rating_true_pred.rename(columns={col_prediction: PREDICTION_COL}, inplace=True)
-
-    return rating_true_pred
+        column_select_true = col_rating
+        column_select_pred = col_prediction
+    return rating_true_pred, column_select_true, column_select_pred
 
 
 @check_column_dtypes
@@ -128,7 +130,7 @@ def rmse(
     col_user=DEFAULT_USER_COL,
     col_item=DEFAULT_ITEM_COL,
     col_rating=DEFAULT_RATING_COL,
-    col_prediction=PREDICTION_COL
+    col_prediction=PREDICTION_COL,
 ):
     """Calculate Root Mean Squared Error
 
@@ -143,13 +145,13 @@ def rmse(
     Returns:
         float: Root mean squared error.
     """
-    rating_true_pred = merge_rating_true_pred(
+    rating_true_pred, column_select_true, column_select_pred = merge_rating_true_pred(
         rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
     )
 
     return np.sqrt(
         mean_squared_error(
-            rating_true_pred[DEFAULT_RATING_COL], rating_true_pred[PREDICTION_COL]
+            rating_true_pred[column_select_true], rating_true_pred[column_select_pred]
         )
     )
 
@@ -176,11 +178,11 @@ def mae(
     Returns:
         float: Mean Absolute Error.
     """
-    rating_true_pred = merge_rating_true_pred(
+        rating_true_pred, column_select_true, column_select_pred = merge_rating_true_pred((
         rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
     )
     return mean_absolute_error(
-        rating_true_pred[DEFAULT_RATING_COL], rating_true_pred[PREDICTION_COL]
+        rating_true_pred[column_select_true], rating_true_pred[column_select_pred]
     )
 
 
@@ -206,11 +208,11 @@ def rsquared(
     Returns:
         float: R squared (min=0, max=1).
     """
-    rating_true_pred = merge_rating_true_pred(
+        rating_true_pred, column_select_true, column_select_pred = merge_rating_true_pred((
         rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
     )
     return r2_score(
-        rating_true_pred[DEFAULT_RATING_COL], rating_true_pred[PREDICTION_COL]
+        rating_true_pred[column_select_true], rating_true_pred[column_select_pred]
     )
 
 
@@ -236,11 +238,11 @@ def exp_var(
     Returns:
         float: Explained variance (min=0, max=1).
     """
-    rating_true_pred = merge_rating_true_pred(
+        rating_true_pred, column_select_true, column_select_pred = merge_rating_true_pred((
         rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
     )
     return explained_variance_score(
-        rating_true_pred[DEFAULT_RATING_COL], rating_true_pred[PREDICTION_COL]
+        rating_true_pred[column_select_true], rating_true_pred[column_select_pred]
     )
 
 
@@ -606,7 +608,7 @@ def auc(
     col_user=DEFAULT_USER_COL,
     col_item=DEFAULT_ITEM_COL,
     col_rating=DEFAULT_RATING_COL,
-    col_prediction=PREDICTION_COL
+    col_prediction=PREDICTION_COL,
 ):
     """
     Calculate the Area-Under-Curve metric for implicit feedback typed
@@ -632,12 +634,12 @@ def auc(
     Return:
         float: auc_score (min=0, max=1).
     """
-    rating_true_pred = merge_rating_true_pred(
+        rating_true_pred, column_select_true, column_select_pred = merge_rating_true_pred((
         rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
     )
     auc_score = roc_auc_score(
         rating_true_pred[DEFAULT_RATING_COL].values,
-        rating_true_pred[PREDICTION_COL].values
+        rating_true_pred[PREDICTION_COL].values,
     )
 
     return auc_score
@@ -650,7 +652,7 @@ def logloss(
     col_user=DEFAULT_USER_COL,
     col_item=DEFAULT_ITEM_COL,
     col_rating=DEFAULT_RATING_COL,
-    col_prediction=PREDICTION_COL
+    col_prediction=PREDICTION_COL,
 ):
     """
     Calculate the logloss metric for implicit feedback typed
@@ -670,13 +672,13 @@ def logloss(
     Return:
         float: log_loss_score (min=-\inf, max=\inf).
     """
-    rating_true_pred = merge_rating_true_pred(
+        rating_true_pred, column_select_true, column_select_pred = merge_rating_true_pred((
         rating_true, rating_pred, col_user, col_item, col_rating, col_prediction
     )
 
     log_loss_score = log_loss(
         rating_true_pred[DEFAULT_RATING_COL].values,
-        rating_true_pred[PREDICTION_COL].values
+        rating_true_pred[PREDICTION_COL].values,
     )
 
     return log_loss_score
@@ -707,5 +709,4 @@ def get_top_k_items(
         .apply(lambda x: x.nlargest(k, col_rating))
         .reset_index(drop=True)
     )
-