… notebooks/02_model/vowpal_wabbit_deep_dive.ipynb