### Description
<!--- Describe your changes in detail -->
<!--- Why is this change required? What problem does it solve? -->
Refactoring and speeding up the rating evaluation

In my experiments the speed up is around half:
```
%timeit rmse(rating_true_big, rating_pred_big) #the original one
1.05 s ± 17.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

%timeit rmse_faster(rating_true_big, rating_pred_big) #the new one
474 ms ± 13.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

The bottleneck was in the function `merge_rating_true_pred` that consisted of 2 operations, a merge and a rename. Curiously, the rename took even more time that the merge. @msalvaris and I just realized that we don't really need the rename so we can pass the name of the true and prediction columns to the sklearn metric

### Related Issues
<!--- If it fixes an open issue, please link to the issue here. -->


### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My code follows the code style of this project, as detailed in our [contribution guidelines](../CONTRIBUTING.md).
- [ ] I have added tests.
- [ ] I have updated the documentation accordingly.